Simplify migration: remove .legacy/.new pattern, consolidate into main files
Per review feedback, the .legacy.tsx/.new.tsx file pattern added unnecessary
complexity. Since styled-components is being removed entirely, there's no need
to maintain backward compatibility wrappers.

Changes:
- Consolidated 6 components: TruncatedText, ShowMyHighlightsStyles,
  HighlightDeleteWrapper, ColorIndicator, Highlights, HighlightListElement
- Moved plain CSS/React implementations from .new.tsx directly into main .tsx
- Deleted all 12 .legacy.tsx and .new.tsx files
- Each component now has simple structure: Component.tsx + Component.css

Updated PLAIN_CSS_MIGRATION_LEARNINGS.md:
- Added "Simplified Migration Strategy" section explaining the change
- Deprecated the .legacy/.new pattern approach
- Updated checklist to reflect in-place migration strategy
- Clarified that .legacy.ts files are only for style utilities, not components

Benefits of simplified approach:
- Fewer files to manage (2 files instead of 4 per component)
- No indirection through re-export shims
- Clearer migration path
- Easier to review
- Matches end goal of removing styled-components entirely

Related: CORE-2282 (PR #3037 Review 6)


Provide id for TruncatedText tests
Provde default class for Highlights

snaps

Fix ColorIndicator style prop undefined in tests

Addresses Review 8 feedback where tests were failing with:
"TypeError: Cannot read property 'passive' of undefined"

Problem:
When ColorIndicator was used as a styled-components component selector
(e.g., `${ColorIndicator}` in ColorFilter.tsx line 104), the `style`
prop was not being passed through correctly, causing it to be undefined
in tests.

Root Cause:
In commit 0eb22f3a, the .legacy.tsx wrapper file was removed as part of
simplifying the migration approach. The legacy file had wrapped
ColorIndicator with `styled(ColorIndicator)` which made it compatible
with styled-components component selectors.

Solution:
Wrap the default export with styled() to restore compatibility with
styled-components component selectors while keeping the plain CSS
implementation. This is a minimal wrapper that:
- Maintains backward compatibility with existing code using component
  selectors (ColorFilter.tsx, ChapterFilter.tsx)
- Allows the style prop to pass through correctly
- Keeps the component's plain CSS styling intact

The wrapped export is only needed because other files still use
styled-components and reference ColorIndicator via component selectors.

Related: CORE-2282, Review 8


Fix ShowMyHighlightsBody to support forwardRef

Addresses Review 7 feedback where ShowMyHighlightsBody was not accepting
refs, causing a TypeScript error in ShowMyHighlights.tsx line 93.

Changes:
- Converted ShowMyHighlightsBody from a regular function component to
  use React.forwardRef pattern
- Now properly forwards refs to the underlying PopupBody component
- Matches the pattern used by PopupBody itself (PopupStyles.tsx:66)

This fixes the TypeScript error:
  Type '{ children: ...; ref: RefObject<HTMLElement>; ... }' is not
  assignable to type 'IntrinsicAttributes & ShowMyHighlightsBodyProps'.
  Property 'ref' does not exist on type '...'

The component now correctly accepts and forwards refs, allowing parent
components like ShowMyHighlights to access the DOM element for scroll
event handling.

Related: CORE-2282
