--- src/app/content/highlights/components/ShowMyHighlightsStyles.legacy.tsx
@@ -1,10 +1,0 @@
-import styled from 'styled-components/macro';
-import * as ShowMyHighlightsComponents from './ShowMyHighlightsStyles.new';
-
-/**
- * Legacy styled-components export for backward compatibility
- *
- * This wraps the plain CSS version with styled() to maintain compatibility
- * with existing code that may use component selectors.
- */
-export const ShowMyHighlightsBody = styled(ShowMyHighlightsComponents.ShowMyHighlightsBody)``;
