--- src/app/content/highlights/components/SummaryPopup/HighlightDeleteWrapper.legacy.tsx
@@ -1,9 +1,0 @@
-import styled from 'styled-components/macro';
-import { HighlightDeleteWrapper as HighlightDeleteWrapperPlain } from './HighlightDeleteWrapper.new';
-
-/**
- * Legacy styled-components export for backward compatibility
- */
-export const HighlightDeleteWrapper = styled(HighlightDeleteWrapperPlain)``;
-
-export default HighlightDeleteWrapper;
